Code formatting, added some styling, deleted unused packages

+++ backend/controllers/invoiceController.js
@@ -4,60 +4,17 @@ const Invoice = require("../models/invoice-model");
 const Customer = require("../models/customer-model");
 const Log = require("../models/log-model");
 
-function generateLog(reqBodyOwner,reqBody,message){
-
+function generateLog(reqBodyOwner, reqBody, message) {
   let stamp = Date();
-  let logdata = {owner:reqBodyOwner,timestamp:stamp,details:reqBody,action:message}
+  let logdata = {
+    owner: reqBodyOwner,
+    timestamp: stamp,
+    details: reqBody,
+    action: message,
+  };
 
-  Log.create(logdata).then(()=>{
-
-  })
+  Log.create(logdata).then(() => {});
 }
-
-// Index: GET all the Invoices
-router.get("/list", (req, res, next) => {
-  Invoice.find({ owner: (req.user._id? req.user._id : req.user.id) })
-    .then((invoices) => {
-      invoices.sort((a, b) => a.customer.localeCompare(b.customer));
-      let data = { invoices: invoices, username: (req.user._id? req.user.username : req.user.displayName) };
-      res.render(`invoiceIndex`, { data });
-    })
-    .catch(next);
-});
-
-router.get("/add", (req, res, next) => {
-  Customer.find({ owner: (req.user._id? req.user._id : req.user.id) })
-    .then((customers) => {
-      customers.sort((a, b) => a.name.localeCompare(b.name));
-      let data = { customers: customers, username: (req.user._id? req.user.username : req.user.displayName) };
-      res.render(`addinvoice`, { data });
-    })
-    .catch(next);
-});
-
-router.get("/:id", (req, res, next) => {
-  Invoice.findById(req.params.id)
-    .then((invoice) => {
-      Customer.find({ name: invoice.customer, owner: (req.user._id? req.user._id : req.user.id) })
-        .then((customer) => {
-          Customer.find({ owner: (req.user._id? req.user._id : req.user.id) }).then((customers) => {
-            customer = customer[0];
-
-            customers.sort((a, b) => a.name.localeCompare(b.name));
-
-            let data = {
-              invoice: invoice,
-              customer: customer,
-              customers: customers,
-              username: (req.user._id? req.user.username : req.user.displayName),
-            };
-            res.render(`invoice`, { data });
-          });
-        })
-        .catch(console.error);
-    })
-    .catch(console.error);
-});
 
 function getTodaysDate() {
   let today = new Date();
@@ -81,41 +38,93 @@ function getTodaysDate() {
   return date;
 }
 
+router.get("/list", (req, res, next) => {
+  Invoice.find({ owner: req.user._id ? req.user._id : req.user.id })
+    .then((invoices) => {
+      invoices.sort((a, b) => a.customer.localeCompare(b.customer));
+      let data = {
+        invoices: invoices,
+        username: req.user._id ? req.user.username : req.user.displayName,
+      };
+      res.render(`invoiceIndex`, { data });
+    })
+    .catch(next);
+});
+
+router.get("/add", (req, res, next) => {
+  Customer.find({ owner: req.user._id ? req.user._id : req.user.id })
+    .then((customers) => {
+      customers.sort((a, b) => a.name.localeCompare(b.name));
+      let data = {
+        customers: customers,
+        username: req.user._id ? req.user.username : req.user.displayName,
+      };
+      res.render(`addinvoice`, { data });
+    })
+    .catch(next);
+});
+
+router.get("/:id", (req, res, next) => {
+  Invoice.findById(req.params.id)
+    .then((invoice) => {
+      Customer.find({
+        name: invoice.customer,
+        owner: req.user._id ? req.user._id : req.user.id,
+      })
+        .then((customer) => {
+          Customer.find({
+            owner: req.user._id ? req.user._id : req.user.id,
+          }).then((customers) => {
+            customer = customer[0];
+
+            customers.sort((a, b) => a.name.localeCompare(b.name));
+
+            let data = {
+              invoice: invoice,
+              customer: customer,
+              customers: customers,
+              username: req.user._id ? req.user.username : req.user.displayName,
+            };
+            res.render(`invoice`, { data });
+          });
+        })
+        .catch(console.error);
+    })
+    .catch(console.error);
+});
+
 router.post("/add", (req, res) => {
-  // getting the current time
   let date = getTodaysDate();
   req.body.createdDate = `${date}`;
   req.body.paid = false;
-  req.body.owner = (req.user._id? req.user._id : req.user.id);
+  req.body.owner = req.user._id ? req.user._id : req.user.id;
   Invoice.create(req.body)
     .then(res.redirect("/invoices/list"))
     .catch(console.error);
-  generateLog(req.body.owner,req.body,"New Invoice created.")
+  generateLog(req.body.owner, req.body, "New Invoice created.");
 });
 
 router.delete("/:invoiceId", (req, res, next) => {
+  let owner = req.user._id ? req.user._id : req.user.id;
+  let details = req.params.invoiceId;
 
-  let owner = (req.user._id? req.user._id : req.user.id)
-  let details = req.params.invoiceId
-  
-  Invoice.findOneAndDelete(
-    { _id: req.params.invoiceId, owner: owner },
-    () => {
-      generateLog(owner,details,"Invoice deleted.")
-      res.redirect("/invoices/list");
-    }
-  ).catch(console.error);
+  Invoice.findOneAndDelete({ _id: req.params.invoiceId, owner: owner }, () => {
+    generateLog(owner, details, "Invoice deleted.");
+    res.redirect("/invoices/list");
+  }).catch(console.error);
 });
 
 router.put("/:invoiceId", (req, res) => {
   Invoice.findOneAndUpdate(
-    { _id: req.params.invoiceId, owner: (req.user._id? req.user._id : req.user.id) },
+    {
+      _id: req.params.invoiceId,
+      owner: req.user._id ? req.user._id : req.user.id,
+    },
     req.body,
     { new: true }
   )
     .then((invoice) => {
-
-      generateLog(invoice.owner,req.body,"Invoice updated ");
+      generateLog(invoice.owner, req.body, "Invoice updated ");
       res.redirect(`/invoices/${req.params.invoiceId}`);
     })
     .catch(console.error);
@@ -123,19 +132,22 @@ router.put("/:invoiceId", (req, res) => {
 
 router.put("/fromcustomer/:invoiceId", (req, res) => {
   Invoice.findOneAndUpdate(
-    { _id: req.params.invoiceId, owner: (req.user._id? req.user._id : req.user.id) },
+    {
+      _id: req.params.invoiceId,
+      owner: req.user._id ? req.user._id : req.user.id,
+    },
     req.body,
     { new: true }
   )
     .then((inv) => {
+      generateLog(inv.owner, req.body, "Invoice updated ");
 
-      generateLog(inv.owner,req.body,"Invoice updated ");
-
-      Customer.find({ name: inv.customer, owner: (req.user._id? req.user._id : req.user.id) })
-        .then((customer) => {
-          res.redirect(`/customers/${customer[0]._id}`);
-        })
-   
+      Customer.find({
+        name: inv.customer,
+        owner: req.user._id ? req.user._id : req.user.id,
+      }).then((customer) => {
+        res.redirect(`/customers/${customer[0]._id}`);
+      });
     })
     .catch(console.error);
 });
